prov/efa: make the current reverse AV lock free
Replace the uthash cur_reverse_av (explicit and implicit) with an efa_av_array,
so the lookup in the cq read is lock free.

The array is indexed by a 32-bit key built from the 16-bit AHN and the 16-bit
QPN. The halves are bit interleaved (a Morton code) rather than packed as
ahn << 16 | qpn: interleaving keeps k-bit halves inside a 2k-bit index, so the
realistic case of small AHNs and small QPNs stays in the array's inline region
and neighboring pairs stay neighbors in the index space. Packing by shifting
would instead place consecutive AHNs 64K indexes apart and burn one chunk per
AHN. x86 gets the interleave as two PDEP instructions when the build targets
BMI2; everything else uses the portable shift/mask spread.

To address the whole 32-bit key space, efa_av_array's max_idx ceiling grows to
UINT32_MAX and max_idx widens to uint64_t.

+++ prov/efa/test/gtest/efa_gtest_av_array_utils.h
@@ -17,9 +17,9 @@ struct efa_av_array;
 /* Create with the default max_idx; NULL on failure. */
 struct efa_av_array *efa_test_av_array_create(void);
 /* Create with a custom max_idx; NULL if the value is rejected. */
-struct efa_av_array *efa_test_av_array_create_max(unsigned max_idx);
+struct efa_av_array *efa_test_av_array_create_max(uint64_t max_idx);
 /* Create with explicit attr fields; NULL if rejected. 0 selects defaults. */
-struct efa_av_array *efa_test_av_array_create_attr(unsigned max_idx,
+struct efa_av_array *efa_test_av_array_create_attr(uint64_t max_idx,
 						   unsigned inline_size,
 						   unsigned chunk_size);
 void efa_test_av_array_destroy(struct efa_av_array *arr);
@@ -37,8 +37,8 @@ int efa_test_av_array_iter_first_hit(struct efa_av_array *arr, void *target);
 
 extern const int efa_test_av_array_inline_size;
 extern const int efa_test_av_array_chunk_size;
-extern const int efa_test_av_array_max_idx_ceiling;
-extern const int efa_test_av_array_default_max_idx;
+extern const uint64_t efa_test_av_array_max_idx_ceiling;
+extern const uint64_t efa_test_av_array_default_max_idx;
 
 #ifdef __cplusplus
 }
